+++ tool/istarcore/fileManager.js
@@ -85,7 +85,7 @@ function saveModel() {
         'actors': [],
         'dependencies': [],
         'links': [],
-        'visualData': {},
+        'display': {},
         'tool': 'pistar.1.0.1',
         'istar': '2.0',
         'saveDate': date,
@@ -93,6 +93,7 @@ function saveModel() {
     };
 
     var toCollapse = [];
+    var vertices = [];
 
     _.each(istar.graph.getElements(), function (element) {
         if (element.isKindOfActor()) {
@@ -120,11 +121,17 @@ function saveModel() {
         if (link.isContributionLink()) {
             linkJSON.label = link.attributes.labels[0].attrs.text.text;
         }
+
+        var vertices = link.get('vertices');
+        if (vertices) {
+            modelJSON.display[link.id] = {vertices: vertices};//add the vertices to the save file
+        }
+
         modelJSON.links.push(linkJSON);
     });
 
     _.each(toCollapse, function (actor) {
-        modelJSON.visualData[actor.id] = {collapsed: true};//add the collapsing information to the save file
+        modelJSON.display[actor.id] = {collapsed: true};//add the collapsing information to the save file
         actor.collapse();//collapses the actor, thus returning it to its original state
     });
 
@@ -159,7 +166,7 @@ function loadModel(inputRaw) {
                     var child = fileManager.addLoadedElement(actor.nodes[j]);
                     if (child) parent.embedNode(child);
                 }
-                if (inputModel.visualData && inputModel.visualData[actor.id]) {
+                if (inputModel.display && inputModel.display[actor.id]) {
                     toCollapse.push(parent);
                 }
             }
@@ -175,6 +182,20 @@ function loadModel(inputRaw) {
                 links[0].on('change:vertices', ui._toggleSmoothness);
                 links[1].on('change:vertices', ui._toggleSmoothness);
 
+                for (j = 0; j < inputModel.links.length; j++) {
+                    linkJSON = inputModel.links[j];
+                    if (linkJSON.target === element.id) {
+                        if (inputModel.display && inputModel.display[linkJSON.id] && inputModel.display[linkJSON.id].vertices) {
+                            links[0].set('vertices', inputModel.display[linkJSON.id].vertices);
+                        }
+                    }
+                    if (linkJSON.source === element.id) {
+                        if (inputModel.display && inputModel.display[linkJSON.id] && inputModel.display[linkJSON.id].vertices) {
+                            links[1].set('vertices', inputModel.display[linkJSON.id].vertices);
+                        }
+                    }
+                }
+
                 dependum.prop('position/x', element.x);
                 dependum.prop('position/y', element.y);
                 // treat as dependum
@@ -182,11 +203,15 @@ function loadModel(inputRaw) {
 
             //create links
             for (i = 0; i < inputModel.links.length; i++) {
-                if (fileManager.isDependencyLink(inputModel.links[i])) {
+                var linkJSON = inputModel.links[i];
+                if (fileManager.isDependencyLink(linkJSON)) {
                     //fileManager.addDependencyLink(inputModel.links[i]);
                 }
                 else {
-                    fileManager.addLoadedLink(inputModel.links[i]);
+                    var newLink = fileManager.addLoadedLink(linkJSON);
+                    if (inputModel.display && inputModel.display[linkJSON.id] && inputModel.display[linkJSON.id].vertices) {
+                        newLink.set('vertices', inputModel.display[linkJSON.id].vertices);
+                    }
                 }
             }